move converter listener to api
