+++ declarative-pipeline-migration-assistant/src/main/java/io/jenkins/plugins/todeclarative/actions/ToDeclarativeAction.java
@@ -10,7 +10,7 @@ import io.jenkins.plugins.todeclarative.converter.api.ConverterRequest;
 import io.jenkins.plugins.todeclarative.converter.api.ConverterResult;
 import io.jenkins.plugins.todeclarative.converter.api.Warning;
 import io.jenkins.plugins.todeclarative.converter.freestyle.FreestyleToDeclarativeConverter;
-import io.jenkins.plugins.todeclarative.listener.DeclarativeConverterListener;
+import io.jenkins.plugins.todeclarative.converter.api.DeclarativeConverterListener;
 import jenkins.model.Jenkins;
 import jenkins.model.TransientActionFactory;
 import org.jenkinsci.plugins.pipeline.modeldefinition.ast.ModelASTPipelineDef;
